+++ tests/main_test.py
@@ -64,6 +64,8 @@ class MainGuiTest(unittest.TestCase):
         # mainapp.init_run()
         mainapp.set_annotation_color_selection("#FF00FF")
         mainapp.run_lobuluses()
+        self.assertLess(0.2, mainapp.evaluation.evaluation_history[0]["Lobulus Border Dice"],
+                        "Lobulus segmentation should have Dice coefficient above some low level")
 
     def test_start_gui_no_exec(self):
         # fn = io3d.datasets.join_path("medical", "orig", "CMU-1.ndpi", get_root=True)