+++ changelog.md
@@ -16,6 +16,7 @@
 * [Added] Processing parameters saved to yaml and json file
 * [Added] Collapsed parameter sub-tree with complex parameters
 * [Added] Algorithm evaluation added
+* [Added] Lobulus border segmentation and Central vein segmentation evaluation
 
 ## 0.4.*